Update linux.sh
add  "-D_MWAITXINTRIN_H_INCLUDED -D_FORCE_INLINES" flags to the compile and update the "rm marvin" command.

+++ marvin/linux.sh
@@ -1,11 +1,15 @@
-rm marvin
+# If marvin file exit, rm marvin will not return error
+file="marvin"
+if [ -f "$file" ]
+then
+    rm marvin
+fi
 
 #g++ -std=c++11 -Wfatal-errors -O3 -o marvin marvin.cpp -I. -I/usr/local/cuda/include -L/usr/local/cuda/lib64 -lcudart -lcublas -lcudnn
 
 export PATH=$PATH:/usr/local/cuda/bin
 
-nvcc -std=c++11 -O3 -o marvin marvin.cu -I. -I/usr/local/cudnn/v3/include/ -I/usr/local/cuda/include -L/usr/local/cudnn/v3/lib64 -L/usr/local/cuda/lib64 -lcudart -lcublas -lcudnn
-
+nvcc -std=c++11 -O3 -o marvin marvin.cu -I. -I/usr/local/cudnn/v3/include/ -I/usr/local/cuda/include -L/usr/local/cudnn/v3/lib64 -L/usr/local/cuda/lib64 -lcudart -lcublas -lcudnn -lcurand -D_MWAITXINTRIN_H_INCLUDED -D_FORCE_INLINES
 
 export LD_LIBRARY_PATH=LD_LIBRARY_PATH:/usr/local/cuda/lib64
